Rewritten the comms subsystem.
Client Comms are now split into a receiver and sender channels - the
receiver channel connects to a new HTTP handler which trickle feeds
data while keeping the TCP connection up. This allows the client to be
blocked on the connection for long periods of time (max poll = 10
minutes) while at the same time, as soon as a new flow is launched,
the client is notified immediately.

Having an instant notification mechanism makes the client extremely
responsive - we can now navigate the VFS in real time!