Support new NTFS APIs.
Refactored the way that device/path are handled in glob and fs
accessors. Now a glob root may be deduced by the accessor by providing
a GetRoot() method. This allows a glob prefix to be considered a
single unit instead of interpreting it as a separated path.

For example the following glob:

\\.\C:\Windows\System32\*.exe

will be interpreted as a glob of \Windows\System32\*.exe from a device
root of \\.\C: