Refactor accessor API to allow root containing path sep.
On windows devices have path separators, but should not be considered
as distinct paths by glob. This change allows an accessor to specify a
root part of the path and begin globbing under that root.

For example, the path:

\\.\c:\Windows\System32

specifies a root of \\.\c: and a path of \Windows\System32

The splitting of paths is done by the accessor's GetRoot() method and
so everything else is unaware of this. Specifically the GUI simply
passes the full path back to the accessor and will render a single
directory with a name of \\.\c: to represent the device.