Added an execve VQL plugin.
This allows arbitrary code execution on the client. Using the output
of local commands is useful - for example reading the output of
ipconfig is easier than calling complex and OS version dependent APIs.

However, allowing arbitrary code execution also increases the risk of
misuse. Velociraptor mitigates this risk by:

1. Logging all execve calls for auditing in the flow logs.
2. If the client configuration prevent_execve is set to true then
   execve is not allowed on the client at all.

This change also adds another fs accessor: 'data'. This makes
arbitrary strings (passed in as the filename) appear as files. This
allows using VQL plugins and functions that normally operate on files
with simple strings. For example the output of execve can be used as
the input for hash, yara scan etc.