Implemented exported files.
Sometimes we need to include a lot of data in an artifact - for
example a huge file containing yara rules. It is more convenient to
include this file instead of embedding it in the artifact itself.

We have a part of the server's VFS named "/exported_files" which
contain just flat files. This commit adds the ability to refer to this
in the artifact collector. The artifact collector will embed the
exported file inside a VQL environment variable so it can be used by
the query. The file may be updated at any time via any
mechanism (e.g. rsync). Each time the artifact collector is run, the
file will be re-read and resent to the client.

Some use cases for this feature:
 - Maintain a list of blacklisted chrome extensions.
 - Maintain a list of bad hashes.
 - Maintain a file of yara rules.