Implemented maximum post size on client communicator.
Previously the client communicator would queue 100 messages in memory
when forming the POST request. This is too large when uploading a
large file because the queue can fill up quickly before it can get
sent. In that case the POST size might be huge (100mb) and will not be
accepted by the server.

This change controls the POST size by limiting it to 5mb, and then
immediately triggering a POST operation while blocking the
executor. This ensures that POST sizes are limited. This also reduces
memory consumption on the endpoint.