Windows service will now wait for config file.
When installing the windows service it will search for a config file
in the same directory as the executable. If the config file is not
present, the service will start but wait for the config file to
appear. When the file appears, the service will start the client and
operate as normal.

This approach allows us to distibute a generic signed Velociraptor MSI
that can be installed via SCCM, while config files can be pushed
separately via GPO.