Added an AppCompatCache parser.
Also fixed opening of large registry keys. When listing registry keys,
a stat response will have a Data field which is a dict containing the
binary value only as long as the value is less than 4kb. For larger
values we need to call read_file() to read the value data into a
larger buffer.