Yara can now accept shorthand rules.
It is tedious to write yara rules for simple cases. Therefore now we
accept a shorthand and generate the rule automatically:

wide:foo,bar,baz

Will generate a rule that matches any of "foo", "bar", "baz" in
utf16. Keywords are split on , and striped of whitespace.

Also fixed bug in file finder to upload only matching files.