Started implementing local ringbuffer cache. (#39)
The local ring buffer is a local cache of outbound messages that are to be sent. The client keeps those in memory or on disk until they are able to be sent. The local buffers allow us to produce messages even when the client is not connected (e.g. offline).

This feature is essential to making client event queries work when offline. Previously the VQL query stopped if the client was unable to upload the rows. Now the query continues as normal, filling the local ring buffer.

The ring buffer is inserted between the executor and the http sender. The sender will attempt to package messages from the ring buffer and deliver them to the server on its own schedule (consisting of max_poll reconnection attempts). As it packages a bunch of messages, it takes a lease on them in the ring buffer. If the messages are delivered correctly, the sender will commit the lease into the ring buffer which removes them from the buffer. All the while the executor is able to insert more messages to the local buffer independently.

When the local buffer is full, the executor is blocked and the query pauses - as is the current behavior.