Store client event queries in writeback.
Client event queries are monitoring queries that are set by the server
for the client to run permanently. The events are updated using the
UpdateEventTable client action.

Currently this is synced via the foreman on every receiver
connections. This is inadequate because if the client starts without
network connectivity it will not run any event monitoring
queries. Since we have a ring buffer now to queue responses - even if
the client is offline, it makes sense to also store the monitoring
queries on the client so we can start monitoring as soon as the client
is started - even if the machine is offline.