Introduced lazy_ntfs accessor.
This experimental accessor is optimized for searching globs. It only
scans the I30 indexes when listing a directory and returns a lazy
object. When the object is accessed for its metadata - only then does
it read the corresponding MFT entry.

The downside of this approach is that the lazy_ntfs accessor is not
aware of ADS stream since it does not inspect the attribute name at
all. However it is much faster than the regular ntfs accessor.