Brought back the elastic plugin. (#96)
The reason it has a huge footprint is due to inefficiently written
client library. Forked client library and make it more efficient (only
export the bulk upload API).

Also added a server side event monitoring artifact which uploads to
elastic the results of any completed flows. This allows the results of
collected artifacts to be examined/visualized in Kibana.