Sanitize artifact YAML to preserve white space in VQL.
YAML has two types of block level scalars. The default one
(which is more intuitive to use) does not preserve white
space. This leads to terrible rendering in the GUI and
elsewhere because the query appears all on the one
line. The user should use the literal scalar (i.e. '- |'
form) but this is a difficult yaml rule to remember and
makes the artifact look terrible.

Therefore we just transform one form into the other in

Previously artifacts had to be written like this or the GUI showed the
VQL on one line:

```
sources:
  - queries:
      - |
        LET completions = SELECT * FROM watch_monitoring(
             artifact="System.Flow.Completion")
        WHERE Flow.FlowContext.artifacts_with_results =~ ArtifactNameRegex
```

and now the whitespace will be preserved in the GUI:

```
sources:
  - queries:
      - LET completions = SELECT * FROM watch_monitoring(
             artifact="System.Flow.Completion")
        WHERE Flow.FlowContext.artifacts_with_results =~ ArtifactNameRegex
```