Refactored GrrMessage to remove client action plugins. (#129)
The GrrMessage protobuf is complex with double encoding due to the
need to have multiple plugins. Since in Velociraptor we have very few
plugins now it makes sense to remove the complexity and just include
all the args in the the same protobuf. This majorly simplifies the
code and speeds things up.