Refactored VFS code to use a service instead of flows. (#137)
VFSRefreshDirectory and VFSDownloadFile are now regular
artifacts. This allows them to be inspected in the GUI in the same way
as other artifacts.

VFSService is a server side service listening for new completions of
these artifacts and adjusting the VFS data structures accordingly.

The VFS now contains references to the real uploaded files in a
VFSDownloadInfo protobuf. This allows the user to re-download a file
into the VFS without trashing the previous version.