Added an append flag to the config repack command.
This allows us to append an external zip file to the Velociraptor
executable while repacking. The added zip file will be covered by the
.rscs section and therefore will be protected by the authenticode
signature.

The standard Golang zip library does not support processing of zip
files with prepended data. The PR also includes a fork of the standard
library which supports this.