API cert and frontend cert can be the same name.
Previously it was assumed that the API gateway and frontend common
name will be different because this is how we generate these
certs. This is confusing though when using externally generated names
so we relax this requirement in this change.

The internal grpc connection will be sent from the frontend name
without a context metadata, and even though it matches the api gw
pinned name, we will fall back to setting the user name from the
common name.