Bugfix: hunt manager keeps an LRU of clients
Due to the interaction with the client side ring buffer foreman
checkins may claim the last hunt id is older than the one actually ran
on the client. This change gets the hunt manager to keep a local LRU
of the hunts each client ran. This prevents multiple hunt launches.