Hunt manager now maintains an index of clients (#191)
Previously we relied on client timestamp alone to tell us when the client participated in a hunt. This is susceptible to race conditions as the client may have sent an enroller message queued up in its ring buffer before updating its last hunt timestamp. This sometimes result in the same flow being scheduled twice or more per client.

We now avoid this problem by maintaining an index in the hunt manager to keep track of which client we issued the hunt for. Checking the index only happens in the hunt manager when the client has indicated it would like to participate in the hunt. Timing here is not critical so it is ok to spend time on the index.

Additionally this PR changes the http client to use the context for timeout instead of relying on the dialer timeout (which was too short at 10 sec).