Added server plugins (#241)
file_store_delete(), enumerate_flow() are used to manage flows on the server. One can now write artifacts to delete old flows in order to manage disk space effectively.

The utf16_encode artifact allows one to encode a string as utf16 - this is used to produce a powershell encoded command for integration with powershell.