Ring buffer file is now OS dependent. (#264)
Also added some checks to ensure the user running the frontend is the
expected user. This tries to remove some common errors of users
installing the deb and then running velociraptor as root. This breaks
the system because running as root will creates files that are not
readable by the low privilege velociraptor user that the service uses.

Now when running from a deb, the frontend will refuse to run as any
other user than the one configured.

Also added syntax test for curated artifacts and fixed the ones that were broken.