Refactored path manipulation. (#267)
Fixes issues with the VFS path.  This refactor introduces a consistent transformation between path components and paths. In many cases a path component can actually contain path separaters within it. For example in NTFS The first component may be \\.\C: , in the registry VFS it is perfectly valid and common for a key name to contain / and a value name  may contain \ as well.

This change ensures that Velociraptor paths are round trip safe - a component list can be serialized to a path and back again with not errors.