Implemented notebooks in the GUI (#275)
Notebooks are a way of post processing the data from velociraptor hunts.

We often need to run several hunts as part of a particular investigation. It is conveniently to collect notes, results and analysis in the same place. This change introduces the concept of notebooks - heavily inspired by Jupyter notebooks (which we can still use with the API).

A user can create a notebook in the Velociraptor GUI then add cells. There are two types of cells:

1. Markdown cells contains markdown (or actually any of the templating constructs available in the reporting).
2. VQL - the VQL cell simply runs the VQL and shows it in a table.
cells can be rearranged or deleted.

NOTE: Cells are calculated currently upon updating the cell content. This means loading should be quick (since the data is stored in the filestore) but updating may timeout. Future improvements can include the GUI polling each cell for completion.

NOTE2 - there is currently only minimal clash detection and no recovery possible. Future improvements will include better support for multiple editors at the same time.