Support Multi-line VQL statements.
Previously in artifact definitions we had a list of queries in each
artifact source. This is hard to manage and not as intuitive. This
change adds a "query" field in a source which allows multiple VQL
statements to be written in one free form text field.

The old "queries" field is still supported, but it is more intuitive
to use the "query" field instead for new artifacts.

Previously:

sources:
  - queries:
     - LET X = SELECT ....
     - SELECT * FROM ....

Now:

sources:
  - query: |
      LET X = SELECT ...
      SELECT * FROM ...