Added ACL enforcement inside artifact definitions. (#287)
This can be used to restrict who can run which artifact.

Refactored the interactive shell to use an artifact which execve() arbitrary command line. Therefore running this artifact is restricted to only those users who have the EXECVE permission. It is now possible to run the shell command using an api client config file.

It is now also possible to encrypt the private key of API client config files. This will ask for a password every time the key is used.