Initial implementation of mysql datastore (#298)
This is the first step towards multi-frontend support.

In this current version the GUI and client comms mostly work. Left to do is to serve the public and notebook directories from the file store.  Also need to tune performance/indexes but this is a good solid start.

A large part of this change is a refactor of the file store interfaces. There are two kinds of objects:
1. FileReader - used to read files, seekable and closable.
2. FileWriter - used to write files. not seekable - can only append to the file, must be closed.

The restriction on the writer append makes implementation much easier and fits well with the Velociraptor data access pattern.