Updated the fs command to be able to use the filestore.
It is essential to be able to easily copy files into and out of the
filestore. With the new filestore being inside mysql it is otherwise
hard to administer it.

This change adds the ability to see inside the filestore to the `fs ls`,
`fs cp` and `fs cat` command.

It is now possible to add items to the public directory like:

```
velociraptor -v fs cp '*.exe'  fs:///public/
```