Brought syslog watcher inline with evtx watcher architecture.
This should fix Issue #300 although I am still not 100% sure how that
crash was possible.

Issue #300 was a send on closed channel but it seems impossible for
the channel to be closed before the context is done and
distributeLine() has a select() between the context and the
channel.

Anyway this new refactor is a bit more obvious and might avoid the
crash.