Refactored FS accessors to use the scope cache. (#310)
File accessors sometimes need to cache things across the life of the
query and they need to take care of proper destruction when the query
is over. Previously accessor caches were tied to context and
destruction occured when the context was done.

However this is specifically buggy because contexts are being
destroyed throughout the query - for example foreach() will create a
subcontext for each iteration. Caches based on this context will
destroy on each iteration.

This behaviour lead to ntfs accessor bug when an ntfs accessor is used
inside a foreach() plugin - the context would be destroyed on each
iteration which caused the underlying device file handle to be closed
before the query was complete.

This change passes the scope to the accessor's New() method. This
allows accessors to create cached inside the scope cache mechanism
which guarantees they will be available throughout the query. The
scope destructor can then be set to clean up when the entire query is
complete.

Additionally this change allows urgent messages to skip the client
side ring buffer. This helps when interacting with a client from the
GUI while there are heavy hunts since the interactive collection
preempts the hunt data.