Report GUI API errors in the GUI. (#312)
This change wraps all GUI APIs in a promise which writes errors on the
error bar. This avoids the need for every api call to do this and
makes error reporting more consistent.

When the GUI loses connectivity or auth is rejected due to expiry this
is now more obvious.

This change also refreshes the csrf token on every get request to
ensure it will never expire. This avoids csrf expiry issues as the
token just rolls over continuously.