Run Artifact Plugin in a new scope. (#317)
When running a sub-artifact using the `Artifact.` plugin, we used to
copy the scope from the calling context. This causes some weird
effects like the sub artifact being able to access values set in
callers scope and leading to scope over-write warnings.

This change creates a new scope for the sub-artifact. We still need to
propagate global variables (like the ACL manager) to the subscope
though.

Refactored access to special scope variables by using scope builder and helpers.