Add root parameter to glob() plugin. (#323)
The root parameter can be used to start the glob operation from a
specific directory. This is useful when we know the exact path of a
directory which may contain glob characters itself. The root is not
expanded itself.

Previously we just contact the path to the glob but this will fail to
match path directories with glob characters.

This change also uses the new version() function to write a backwards
comatible artifact for older clients.

Fixes: #322