Refactor filestore to add in-memory queuing  (#325)
Previously queuing was done by writing journal files on the filestore. This was done as a way to supporting multi-process queuing.  With the introduction of the mysql backend it became obvious that we can use properties of the backend to help with more efficient queuing and event delivery. We now have a choice:

1. If the number of endpoints is not too large (say <10k) then we can keep most of the queuing in memory and this makes the server very fast and efficient.
2. For larger deployment sizes we now use the mysql backend and therefore we can use a different and more scalable architecture for queuing. It may not be as quick as in-memory queuing but it should be more scalable.

This change moves the event queuing in memory and refactores the filestore to support specific queuing architectures.  The file based filestore no longer writes journal files and instead queues messages in memory to all listeners. Therefore event notifications are also much faster and low latency.