Do not send cancellation messages for unknown flows. (#329)
If the server receives a message from the client referring to an
unknown flow, it will issue a cancellation message. The client will
attempt to cancel the message and log this to the server.

The server will then send another cancellation and this will lead to
ping ponging.

This change ensures that the server does not send cancellation
messages in response to logs or status messages.

To be extra careful we ensure that clients only respond to the first
cancellation message for a flow, then ignore the rest.