Artifact collector can now produce a HTML report.
Reports are produced by evaluating the templates in each collected
artifact's report section. We hook the source() plugin to make sure
the results are read from the produced zip file.

The reports allow one to add notes and interpretation hints to the
collected artifacts.

Currently this is only supported in the standalone collector, but we
should also provide a GUI feature to allow these reports to be added
to a notebook.