Fixed bug in lazy ntfs. (#494)
The lazy_ntfs accessor is faster because it only looks at $I30
records.  The lazy_ntfs accessor was calling record.IsValid() to make
sure the records are valid. However in practice $I30 records are not
always fully populated in particular $LogFile has wrong timestamps.

This change does not force this check any more.