Implemented a client KillKillKill message.
This causes clients to be hard killed and refreshed. It will remove
the ring buffer to ensure the client is reachable again.

Also added extra safety to ensure the client executor loop is always
active. If any part of the loop exits - due to bug or exception, the
client will immediately suicide. The service control manager should
restart the client and bring it back to an operational state.