Automatically upgrade tools when server version changes. (#552)
Tools now automatically upgrade when server version is upgraded. This
is especially important with the offline collector tracking the latest
Github releases.

It is now possible to specify a Github project for a tool instead of a
URL. This allows Velociraptor to query the github API for the latest
release from that tool.

Also implemented better coloring support.