Client monitoring bugfix.
When scheduling multiple event artifact on the client we need to split
each artifact into its own VQLCollectorArgs included in a different
VQLEventTable.Event member.

This is because each Event is run in its own goroutine on the
client (therefore in parallel) which is essential for blocking event
queries.

This behavior is exactly the opposite of scheduling normal artifact
collection, where all the artifacts are compiled into the same
VQLCollectorArgs - because the client collects all artifacts in the
same scope instance in series.