Update client monitoring tables when relevant artifacts are modified
When an event artifact is modified through the GUI we need to
recompile the client monitoring tables, or we risk caching stale data.

This change broadcasts artifact modification events to the
Server.Internal.ArtifactModification artifact. The client event
manager then listens to changes in this artifact and rebuilds the
tables (and increments versions) when the artifact changes.