Added direct download link to uploaded files. (#558)
Sometimes it is such a hastle to prepare a zip export file from a flow
just to get at one file. This change adds direct link to the uploaded
files tab to each file so it can be downloaded individually.

This is very helpful for the offline collector builder which is just an 
artifact that uploads the binary to the server. You can now just download 
the prepared binary immediately from the uploaded files tab.