Implement row limits on collections (#602)
This change keeps track of the number of rows returned from the client
in each VQLResponse inside the collection context. When the number of
returned rows exceeds the limit, the collection is cancelled
automatically.

A similar limit is also implemented using the TotalUploadedBytes
metric. When the collection collects more than this many bytes, the
collection is also cancelled.

Separately from this an optimization is implemented where clients
encode their response into JSONL format. This allows the server to
directly copy the client's responses to the end of the result queue
within needing to decode and encode it again. This improves collection
efficiency a lot - especially for collections that return a lot of rows.