Solidify the rules around when tools can be upgraded.
Tool definitions are upgraded automatically:

1. When an artifact is added and the tool is not already known.
2. Then an admin forces the tool to be updated

In the first case, the artifact may have a tool definition which
includes where the upstream url of the tool is and a hint about the
serving policy (serve locally). The artifact definition is only a
suggestion and multiple artifacts may define the same tool with
potentially conflicting data.

This PR implements the idea of a better/worse tool definition. A
better definition is one which has enough information for us to
actually fetch the tool (e.g. URL, github project).

An artifact may simply declare that its using the tool but not declare
where it should be served from - instead relying on another artifact
to specify more information.

Therefore artifacts are only upgraded opportunistically when they have
more information than previous definition.

Admin specified artifacts are always respected - an artifact
definition will never overwrite an admin specified definition. At the
same time an admin is able to override the previous definition.