Added a cache VQL plugin which stores data for a time.
The cache function produces a scope lifetime cache. It takes 2
parameters, a func and a key. The key is used to index into the cache,
while the func will be evaluated when the cache misses.

The following is an example of how to use the cache:

```
  LET Foo(X) = if(condition=log(message=format(format="I Ran with %v", args=X)),
                    then=X + 5)
  SELECT cache(func=Foo(X=5), key=5),
           cache(func=Foo(X=10), key=5),
           cache(func=Foo(X=10), key=10)
  FROM scope()
```

This will produce a row with (10, 10, 15) - the middle item is cached
because its key is the same as the first item.