Generate certificates with SAN instead of CommonName. (#648)
Go 1.15 broke support for the CommonName field in the certificate
subjects. Velociraptor used the CommonName to encode the server's name
for additional checking (essentially pinning the self signed certs).

This change uses the DNSNames array of the SAN field in the
certificate. Note that it is **not** used as a dns name at all since
it is used to authenticate API connections who have no concept of a
dns name.

Fixes: #648