Added VFS recursive download button in the GUI
It is now possible to recursively download the files in a
directory. This can tranfer huge amount of data! However we now have
download limits so the default is 1gb.

Also added a delete flow and made the old delete flow an archive
flow. This new button actually deletes the data freeing up disk
space (User needs to be admin to use it).

Fixed redirect to go directly to the app.