Offline Collector cloud upload test before collecting.
Sometimes the cloud upload fails due to factors beyond our
control (e.g. clock skew or invalid creds). In that case it is
annoying to only find out about it after collecting a ton of
data. This change uploads a small file first and then aborts the
collecting if the file can not be uploaded before collecting all the
artifacts.

This change also uploads the collection log file to the cloud.