Implemented a TimedResultSetReader (#731)
A TimedResultSetReader is a paging aware, seamless result reader for
monitoring artifacts. Callers may instantiated it over the time range
of interest then page over it efficiently.

This allows us now to implement infinite paging on the event
monitoring screen.

* Refactor Windows.Forensics.LocalHashes.Usn to deduplicate.

The USN Journal has multiple events for the same file change. This
commit merges all modifications occuring within the same interval so
we only need to hash it once.