Shell screen should display shell collections launched normally. (#793)
The format of the CollectArtifact request API was changed to allow
multiple artifacts to run in separate scopes. The system will deal
with both formats (old and new) but the shell screen launched the
artifact using the old format and displayed it only using the old
format. This means that you could see the shell commands issued
through the shell screen but it was not able to properly display
collections issued from the collect artifacts page (which used the new
format).