Do not expose the server_config to VQL
For server side VQL, we previously added the config object to the VQL
environment but this exposes sensitive information to notebook users
who may not have the required permissions.

This change stores the config object in the query cache - this is not
accessible to VQL directly, but VQL plugins can still get it when they
need it.