Add Windows.ActiveDirectory.BloodHound (#840)
This artifact allows deployment of the BloodHound collection tool Sharphound.

BloodHound is a popular Active Directory Assessment tool that uses graph
theory to reveal the hidden and often unintended relationships. It can also be
used to identify and eliminate potentially risky domain configuration.

The Sharphound collection is in json format and upload to the server for
additional processing.

RemovePayload - Due to potential malicious use of this tool I have also
included an option to remove payload execution.

NOTE: Do not run this artifact as an unrestricted hunt. General recommendation
is to run this artifact on only a handful of machines in a typical domain,
then deduplicate output.

author: Matt Green - @mgreen27

reference:

https://github.com/BloodHoundAD/BloodHound
https://github.com/chryzsh/awesome-bloodhound