Implemented multithreaded Zip offline collector
The offline collector uses zip for storing the results, but in the
default Golang implementation the zip writer is single threaded and
can not utilize more than one CPU. This change allows the writer to
run multiple workers and compress multiple files at once.

It seems to lead to acquisition times reducing by a factor of 3-4
approx.