Implemented file buffers for directory queue manager. (#876)
Previously, the queue manager was in memory only, pushing events
through memory channels to consumers. This means if a consumer was too
slow, events were accumulating in memory until the channel was
full. At this point, the back pressure on the senders would cause
server lockup since queues are often read in the critical path.

This change creates an overflow buffer file for each queue
listener. If the listener is not available to read from the channel
immediately, the queue manager will write the data to a file, and
return immediately to the writer.

Therefore it is no longer possible to block on queue writes which
ensures stability under high load.