Pass types to dependent artifacts in request.
This is needed to properly convert default types for dependencies.

When calling the client we attach a distilled version of the artifact
definition of dependent artifacts (i.e. those referenced from within
the VQL). This allows the client to resolve dependencies at runtime.

While args passed directly to dependent args have the correct type
already, defaults are always strings and need to be converted. In
order to do that correctly the client needs to know about their types.

NOTE: This might cause erros on older clients - usually not critical
errors. This happens because current artifacts assume parameters are
already converted to their correct types and if the client is not
doing this correctly they will fail.