Refactor Raw Registry accessor to use paged reader. (#925)
Fixes a file handle leak in the raw registry parser which would cause
errors in mounting hives for the windows profile service.

Fixes #924

Using ntfs accessor in raw_reg artifacts allows us to avoid the file
handle leak issue. This fixes this behavior on old clients which do
not have the fix yet.