Notifications are throttled now. Refactor VQLCollectorArgs compiler. (#958)
This change controls the rate of notifications in the notifications
service. Usually when a new hunt is created, the notifier notifies all
the clients immediately so they can re-connect and receive the next
hunt. On very large deployments this causes a bottleneck and a slow
down of the server.

To address this issue we
1. Limit the rate of notifications to ensure not all clients are
   reconnecting at the same time.
2. Limit the total connection rate to ensure the server is not
   overwhelmed. This simply returns 500 when the connection rate is
   exceeded.

Additionally this PR changes the way tools are passed to the
VQLCollectorArgs when the artifact is compiled. Previously tool
information was passed in the global artifact's
env (e.g. Tool_XXX_HASH), and we relies on the previous behaviour that
all dependent artifacts shared the root scope.

However in recent releases dependent artifacts are running in an
isolated scope and so can not see the tool information. This change
sets the tool information as a parameter in each dependent artifact's
definition to ensure that tool is properly resoved on the client.