Page hunt list after sorting all the hunts. (#959)
Fixed: #953

Previously the hunt list was paged during collecting the hunt objects
from the hunt dispatcher. However the order of traversal in the hunt
dispatcher is non-deterministic so this paging resulted in unstable
hunts list.

This PR fixes that by paging only after the hunt list is sorted by
creating time.