Preconditions are now supported by clients. (#977)
Previously preconditions were implemented as VQL lazy queries. This
works fine as long as there are no materialized queries in the
artifact sources. If there are, the query will be materialized even if
the precondition prevents it from running.

This PR makes clients support precondition natively. The server will
declare the precondition in the request and the client will not
execute any of the VQL if the precondition fails.