Deduplicate event query table when no change is detected. (#1001)
The server may refresh the client's event table when no change is
actually needed (for example when the client's labels change but these
do not affect the label group assignment).

In this case the client may restart its event queries needlessly. This
change ensures that the queries are not refreshed if they didnt
actually change.