Allow NTFS cluster cache size to be specified in VQL. (#1042)
Setting VQL scope variable NTFS_CACHE_SIZE will allow the NTFS cluster
cache to be this large. This reduces disk IO in favor of higher memory
usage. The default cache size is 100 clusters which results in a lot
of cache misses on the typical MFT (but only consumes 800kb).

Also automatically consume BOM for parse_csv() - the BOM is an
aBOMination and often set by Windows software (see
https://pkg.go.dev/github.com/spkg/bom)