Hunt manager can add clients to hunt after the hunt is started
Previously if a client was labeled after the hunt started, the hunt
did not assign on this client. This is not intuitive, as the hunt is
currently active and it is targetting a label which the client
receives.

This change allows the hunt manager to pick up any clients that were
labeled and retry the hunt conditions on those clients.

This allows a hunt to be created targeting a
label (e.g. "Compromised"), then slowly adding clients to this hunt
via label as needed.