Added GUI to quarantine a host. (#1166)
Also added a monitoring artifact that will reload the ipsec policy
periodically on the client. This ensures the policy is loaded even
after a reboot or when the network is changed.

Quarantined hosts are managed via the Quarantine label. Adding the
label will automatically quarantine the host and removing the label
will unquarantine the host. This requires having the
Windows.Remediation.QuarantineMonitor artifact enabled on the
Quarantine label.